# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"uuid": {
						"type": "string",
						"example": "123e4567-e89b-12d3-a456-426614174000"
					},
					"username": {
						"type": "string",
						"example": "admin"
					},
					"email": {
						"type": "string",
						"example": "admin@example.com"
					},
					"displayName": {
						"type": "string",
						"nullable": true,
						"example": "John Doe"
					},
					"role": {
						"type": "string",
						"enum": [
							"USER",
							"MOD",
							"ADMIN"
						],
						"example": "USER"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-01T00:00:00.000Z"
					}
				},
				"required": [
					"uuid",
					"username",
					"email",
					"role",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Note": {
				"properties": {
					"uuid": {
						"type": "string",
						"example": "123e4567-e89b-12d3-a456-426614174000"
					},
					"title": {
						"type": "string",
						"example": "My first note"
					},
					"content": {
						"type": "string",
						"example": "This is the content of my first note."
					},
					"authorId": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-01T00:00:00.000Z"
					}
				},
				"required": [
					"uuid",
					"title",
					"content",
					"authorId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NoteOnUser": {
				"properties": {
					"noteId": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"userId": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"assignedById": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"assignedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-01T00:00:00.000Z"
					},
					"role": {
						"type": "string",
						"enum": [
							"USER",
							"MOD",
							"ADMIN"
						],
						"example": "USER"
					},
					"note": {
						"$ref": "#/components/schemas/Note"
					}
				},
				"required": [
					"noteId",
					"userId",
					"assignedById",
					"assignedAt",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NoteCreateRequest": {
				"properties": {
					"title": {
						"type": "string",
						"example": "My first note"
					},
					"content": {
						"type": "string",
						"example": "This is the content of my first note."
					},
					"authorId": {
						"type": "number",
						"format": "double",
						"example": 1
					}
				},
				"required": [
					"title",
					"content",
					"authorId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NoteOnUserCreateRequest": {
				"properties": {
					"noteId": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"userId": {
						"type": "number",
						"format": "double",
						"example": 1
					}
				},
				"required": [
					"noteId",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterRequest": {
				"properties": {
					"username": {
						"type": "string",
						"example": "newuser",
						"minLength": 3
					},
					"password": {
						"type": "string",
						"example": "secret123",
						"minLength": 6
					},
					"email": {
						"type": "string",
						"example": "user@example.com",
						"format": "email",
						"pattern": "^(.+)@(.+)\\.(.+)$"
					},
					"displayName": {
						"type": "string",
						"nullable": true,
						"example": "John Doe"
					}
				},
				"required": [
					"username",
					"password",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthResponse": {
				"properties": {
					"accessToken": {
						"type": "string",
						"example": "eyJhbGciOiJI..."
					},
					"accessTokenExpiresOn": {
						"type": "number",
						"format": "double",
						"example": 1772719299
					},
					"refreshToken": {
						"type": "string",
						"example": "d1f2e3c4b5a6..."
					},
					"refreshTokenExpiresOn": {
						"type": "number",
						"format": "double",
						"example": 1773324099
					}
				},
				"required": [
					"accessToken",
					"accessTokenExpiresOn",
					"refreshToken",
					"refreshTokenExpiresOn"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"username": {
						"type": "string",
						"example": "admin",
						"minLength": 1
					},
					"password": {
						"type": "string",
						"example": "admin",
						"minLength": 1
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshRequest": {
				"properties": {
					"refreshToken": {
						"type": "string",
						"example": "d1f2e3c4b5a6...",
						"minLength": 1
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "insane-node-rest-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "lucasdee"
		}
	},
	"paths": {
		"/users/me": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/notes": {
			"get": {
				"operationId": "ListMyNotes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/NoteOnUser"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Notes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/note/create": {
			"post": {
				"operationId": "CreateNote",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Note"
								}
							}
						}
					}
				},
				"tags": [
					"Note"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NoteCreateRequest"
							}
						}
					}
				}
			}
		},
		"/note/assign": {
			"post": {
				"operationId": "AssignNoteToUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoteOnUser"
								}
							}
						}
					}
				},
				"tags": [
					"Note"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NoteOnUserCreateRequest"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequest"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshRequest"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshRequest"
							}
						}
					}
				}
			}
		},
		"/auth/revoke-expired-tokens": {
			"post": {
				"operationId": "RevokeExpiredTokens",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}